requirements.txt
Django==3.2.5
djangorestframework==3.12.4
python-dotenv==0.18.0